function_app/requirements.txt
azure-functions
twilio==9.*
openai>=1.30.0
azure-identity
azure-keyvault-secrets
azure-cognitiveservices-speech
azure-storage-blob
